refactor: post component 위치 수정
