--- src/features/posts/index.ts
@@ -1,2 +1,0 @@
-export { WriteButton } from './ui/write-button';
-export * from './api/mutations';
